Moved location of absorbDirective func
